move types to its own file
